[Nexters/covid-letter-web#47] ✨ 비로그인 접근 시 page 네이밍을 /direct 로 변경. - Envelope 공통컴포넌트와 구분
