+++ src/pages/api/letters/[encryptedId].ts
@@ -23,7 +23,7 @@ export default async function handler(
             result: {
                 ...letter,
                 createdDate: new Date().toDateString(), //현재 BE 테스트데이터에 createdDate 가 null 로 되어있어 임시용도
-                state: LetterState.SEND, //todo remove Test 용도
+                name: '최인혁', //todo remove 추후 서버 수정 시 삭제할 것
             },
         })
